Refactor Entity classes and move interfaces to separate directory
This commit performs two primary actions. Firstly, it refactors the Entity classes in the Domain layer, modifying the implementation of the IsTransient method and reorganizing the operator methods. Secondly, it moves all interface definitions from the root Domain/src directory to a new Domain/src/Interfaces directory for better organization and easier navigation.