Update signing scripts for Webserver and Webmaster
This commit updates the signing scripts in both Webserver and Webmaster. Specifically, it removes the `-tl` option from `dotnet clean` and `dotnet pack` commands which helps in making the commands more generic. In Webserver/sign.ps1, a comment has been added for codesign command and redundant signing verbosity options in `dotnet nuget sign` are removed for better readability and maintenance.