Update signing script to consider directory and include timestamp
The signing process in the script has been adjusted in two ways. It now checks the directory for files to sign, instead of just file names. Additionally, the command used for signing is now appending a timestamp from Digicert. This ensures all files in the Release directory get properly signed with a trusted timestamp.