Update version prefixes and formatting settings
The version prefixes for the Hosting, Cryptography, and Tools projects have been incremented as part of the update process. Additionally, the indentation size set in the .editorconfig for .props and .targets files has been changed to 2 for consistency. In the sign.ps1 script, lines for Cryptography, Hosting, and Tools have been uncommented. The script is now signed with these modules.