Refactor HasDuplicates method and add tests
The HasDuplicates method in EnumerableExtensions has been refactored for efficiency. Instead of using Any() with a new HashSet<TProp>() in every iteration, we now create the HashSet once and iterate through the list separately. A new test file, EnumerableExtensionsTests.cs, was also created to validate the functionality of the modified method as well as other methods in EnumerableExtensions.