Refactor Enum extensions and fix processor detection
The Enum extensions were refactored for cleanliness and better readability, reducing unnecessary and confusing code structures. Additionally, the Processor detection method in the PlatformService was corrected, changing the return value from Processor.x86 to Processor.x64 when a PowerPC agent is detected. The test method 'ToStringInvariantFlag' in EnumExtensionsTests was commented out.