Add JetBrains.Annotations package and NowDateTimeOffsetGenerator test class
The JetBrains.Annotations package has been added to the EntityFramework tests, with references updated in Usings.cs and the test project file. Furthermore, the NowDateTimeOffsetGeneratorTests class has been introduced, which tests the functionality of the NowDateTimeOffsetGenerator class. This will strengthen the testing infrastructure of DateTimeOffset value generation mechanisms within the project.